Refactor cookbook structure to allow both fullscreen and drawer examples
