+++ apps/cookbook/src/app/examples/modal-experimental-example/modal-experimental-example.module.ts
@@ -8,7 +8,7 @@ import { KirbyModule } from '@kirbydesign/designsystem';
 import { KirbyModalModule } from '@kirbydesign/designsystem/modal/experimental';
 import { ExamplesSharedModule } from '../examples.shared.module';
 
-import { FullscreenModalExperimentalExampleComponent } from './fullscreen/fullscreen-experimental-example.component';
+import { FullscreenModalExperimentalExampleComponent } from './modal/fullscreen/fullscreen-experimental-example.component';
 import { ModalControllerExperimentalExampleComponent } from './controller/modal-controller-experimental-example.component';
 import { ModalRoutingExperimentalExampleComponent } from './modal-routing/modal-routing-experimental-example.component';
 import { ModalRoutingExperimentalExamplePage1Component } from './modal-routing/modal-routing-experimental-example-page1.component';
