+++ apps/cookbook/src/app/examples/modal-experimental-example/modal/fullscreen/fullscreen-experimental-example.component.scss
